requirements.txt
Flask==2.0.1
tensorflow==2.16.1
Pillow==9.0.1
numpy>=1.26.0,<2.0.0 